+++ src/PjbServer/Tools/Exception/FilePermissionException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace PjbServer\Tools\Exception;
+
+class FilePermissionException extends RuntimeException implements ExceptionInterface
+{
+}