+++ src/PjbServer/Tools/Exception/PortUnavailableException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace PjbServer\Tools\Exception;
+
+class PortUnavailableException extends RuntimeException implements ExceptionInterface
+{
+}