+++ src/PjbServer/Tools/Exception/ServerControlExceptionInterface.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace PjbServer\Tools\Exception;
+
+interface ServerControlExceptionInterface extends ExceptionInterface
+{
+}