+++ src/PjbServer/Tools/Exception/StartFailedException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace PjbServer\Tools\Exception;
+
+class StartFailedException extends RuntimeException  implements ServerControlExceptionInterface
+{
+}