+++ src/PjbServer/Tools/Exception/StopFailedException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace PjbServer\Tools\Exception;
+
+class StopFailedException extends RuntimeException  implements ServerControlExceptionInterface
+{
+}