+++ src/PjbServer/Tools/Exception/UnsupportedSystemException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace PjbServer\Tools\Exception;
+
+class UnsupportedSystemException extends RuntimeException  implements ExceptionInterface
+{
+}